requirements.txt
python-dotenv
requests
pymongo
pytest
pylint
